+++ src/components/index.ts
@@ -5,4 +5,3 @@ export { default as TickerRow } from "./TickerRow";
 export { default as YuzuHeader } from "./YuzuHeader/YuzuHeader";
 export { default as MarketHeader } from "./MarketHeader/MarketHeader";
 export { default as WatchList } from "./WatchList/WatchList";
-export { default as YuzuHome } from "./YuzuHome/YuzuHome";
