+++ src/routes/index.ts
@@ -1,0 +1,3 @@
+export { default as Details } from "./Details";
+export { default as Home } from "./Home";
+export { default as Root } from "./Root";
